Update 1st Page Code.html
For this commit, I coded the layout of the 1st topic webpage. There is a background image, a logo at the upper left corner, 2 images that are related to hallyu, as well as 3 paragraphs composed of about 90 words (Lorem Ipsum for now because I assigned my partner to compose the texts). There are also 5 links that lead the user back to the home page, and the 4 other topic web pages.